Create Stochastic Algorithm .py
Stochastic Hill climbing is an optimization algorithm.

It makes use of randomness as part of the search process. This makes the algorithm appropriate for nonlinear objective functions where other local search algorithms do not operate well.

It is also a local search algorithm, meaning that it modifies a single solution and searches the relatively local area of the search space until the local optima is located. This means that it is appropriate on unimodal optimization problems or for use after the application of a global optimization algorithm.

In this tutorial, you will discover the hill climbing optimization algorithm for function optimization

After completing this tutorial, you will know:

Hill climbing is a stochastic local search algorithm for function optimization.
How to implement the hill climbing algorithm from scratch in Python.
How to apply the hill climbing algorithm and inspect the results of the algorithm.